Server core data model
Created a data model to match the systems requirements. A total of 10
POJO's and 15 database tables. Note that comments for the model need to
be added. Close #68

Updated Spring to 1.1.8 Close #75

Removed database definitions and import statements for the old User,
Treasure and Trek objects. This means that the current test database
will output no data. The objects themselves are still in the code
however and will be removed when the new data model is completed. Close
#76