AuthenticationService completed #100
The logic for new player creation is now present. Note that
/registration is not yet fully functional. 

ServerResponse renamed to RegistrationResponse and added a field for id.
Created a lot of tests for user and province repositories.
AuthenticationService is 100% covered with tests. 

New method in ProvinceRepository for finding a province with specific
coordinates. 

Created constants for player sid lenght and player starting units. 

Removed Spring boot actuator.