Unit movement creation close #140
Movements are now created and stored in the database. Untested for
multiple BeginMovementDTOs.

Refactored playerStrength calculation to be a part of the Player class.
Also the same for unit counting.

New class for holding game logic. The content is untested.

Stuff thats bad: date to return to client is currently arbritrary. No
calculation when to reach the destination. Also the exceprtion handling
could use some work.