Unit transfer methods. Close #144
At the end of a movement, units will be added to province units. Works
with home province and normal owned province. 

Started working on a MovementWorker that would run in the background and
complete the movement automatically. Unfortunately there are problems
with Hibernate session management, so unit movement is NOT fully
functional yet. #141

This has been completed long ago. Close #139

Updated server to 0.5.0